rename: recruit 관련 validator 위치 수정
